Renaming the user datastore adapter to platform module
This will be useful to wrap the datastore, submitter and future
filestore into one module that represents the platform
